Je code marche pour des points 2D et 3D maintenat il faut faire la conversion entre les points.

+++ Prosit_3/Point2D.h
@@ -10,7 +10,7 @@ class Point2D
 	float x;
 	float y;
 	float z;
-	float A;
+	double A;
 
 public:
 	void Point();
